move WSClientTestCase to test_utils
